Split controller configuration apply flow
